% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epalmer_test.R
\name{epalmer_test}
\alias{epalmer_test}
\title{A Pretty Awesome Test Function}
\usage{
epalmer_test(adj = "Awesome")
}
\arguments{
\item{adj}{How would you describe Elliot? Defaults to Awesome.}
}
\description{
This function allows you to show your
}
\examples{
epalmer_test(adj = "Amazing")
#Elliot is Amazing!!!
}
\keyword{elliot}
