% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_commas.R
\name{format_commas}
\alias{format_commas}
\title{format_commas}
\usage{
format_commas(numbers)
}
\arguments{
\item{numbers}{a vector of numbers}
}
\description{
format numbers as string with commas
}
\examples{
x <- c(100,1000,10000,10000)
format_commas(x)
}
\keyword{format}
