% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_parentheses.R
\name{format_parentheses}
\alias{format_parentheses}
\title{format_parentheses}
\usage{
format_parentheses(numbers)
}
\arguments{
\item{numbers}{a vector of numbers}
}
\description{
format negative numbers with parentheses
}
\examples{
x <- c(100,-1000,10000,-0.10)
format_parentheses(x)
}
\keyword{format}
