% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percent.R
\name{format_percent}
\alias{format_percent}
\title{format_percent}
\usage{
format_percent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{a vector of decimals}

\item{digits}{the number of digits to show in the}
}
\description{
format decimals as a percentage
}
\examples{
x <- c(.01,.02,.99,-.19)
format_percent(x)
}
\keyword{format}
